refactoring of all major components into /app folder; reorganized modules into using index file to import
