+++ client/src/components/app/About.js
@@ -1,9 +1,9 @@
 import React from "react";
-import "../styles/styles.css";
-import { colors } from "../styles/theme.js";
+import "../../styles/styles.css";
+import { colors } from "../../styles/theme.js";
 import Typist from "react-typist";
 import TypistLoop from "react-typist-loop";
-import PhoneFrame from "./cards/PhoneFrame";
+import { PhoneFrame } from "../frames";
 
 export default class AboutComponent extends React.Component {
   render() {
