+++ client/src/components/app/index.ts
@@ -1,0 +1,4 @@
+import Faq from "./Faq";
+import Members from "./Members";
+
+export { Faq, Members };
