+++ client/src/components/frames/index.ts
@@ -1,4 +1,4 @@
 import PhoneFrame from "./PhoneFrame";
 import PolaroidFrame from "./PolaroidFrame";
 
-export default [PhoneFrame, PolaroidFrame];
+export { PhoneFrame, PolaroidFrame };
