+++ client/src/index.js
@@ -2,10 +2,9 @@ import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
 import Home from "./home";
-import Members from "./components/Members";
-import Faq from "./components/Faq";
-import Recruitment from "./components/Recruitment";
-import NavBar from "./components/NavBar";
+import { Faq, Members } from "./components/app";
+import Recruitment from "./components/app/Recruitment";
+import NavBar from "./components/app/NavBar";
 import * as serviceWorker from "./serviceWorker";
 import { Location, Router } from "@reach/router";
 import "bootstrap/dist/css/bootstrap.min.css";
